+++ index.commonjs.js
@@ -1,4 +1,6 @@
 exports.download = require('./commonjs/download').default
 exports.generate = require('./commonjs/generate').default
-exports.compress = require('./commonjs/compress').default
+exports.minify = require('./commonjs/minify').default
 exports.version = require('./commonjs/version').default
+// Deprecated export: `compress` (now called `minify`).
+exports.compress = require('./commonjs/minify').default