+++ index.js
@@ -1,4 +1,6 @@
 export { default as download } from './modules/download'
 export { default as generate } from './modules/generate'
-export { default as compress } from './modules/compress'
+export { default as minify } from './modules/minify'
 export { default as version } from './modules/version'
+// Deprecated export: `compress` (now called `minify`).
+export { default as compress } from './modules/minify'